various fixes and improvements, mostly for EFS sync
EFS data is sometimes successfully received and the modem stays online
because the reset after the sync works (even if the EFS sync itself
fails). EFS data is not yet written to the partition. This needs to be
implemented.
When EFS data is read, timeouts are now checked and the data is read
in chunks which mirrors the behaviour of the kickstart tool. Strangely
enough, occasionally after receiving the first package, that is below
the maximum size, a timeout happens and the complete data is never
received during all runs. Either the modem is not sending the complete
data for whatever reason or there is still something wrong with the
command chain.

Signed-off-by: Wolfgang Wiedmeyer <wolfgit@wiedmeyer.de>